Add component for now reading
